+++ app/src/main/java/com/boostcamp/dreampicker/presentation/main/SplashActivity.java
@@ -6,6 +6,7 @@ import com.boostcamp.dreampicker.R;
 import com.boostcamp.dreampicker.data.common.FirebaseManager;
 import com.boostcamp.dreampicker.databinding.ActivitySplashBinding;
 import com.boostcamp.dreampicker.presentation.BaseActivity;
+import com.boostcamp.dreampicker.presentation.main.login.LogInActivity;
 
 import java.util.concurrent.TimeUnit;
